+++ src/pybind/py_e2sarDP.cpp
@@ -104,6 +104,12 @@ void init_e2sarDP_segmenter(py::module_ &m) {
         .def_readwrite("numSendSockets", &Segmenter::SegmenterFlags::numSendSockets)
         .def_readwrite("sndSocketBufSize", &Segmenter::SegmenterFlags::sndSocketBufSize)
         .def_readwrite("rateGbps", &Segmenter::SegmenterFlags::rateGbps)
+        .def_readwrite("warmUpMs", &Segmenter::SegmenterFlags::warmUpMs)
+        .def_readwrite("eventQueueSize", &Segmenter::SegmenterFlags::eventQueueSize)
+        .def_readwrite("smooth", &Segmenter::SegmenterFlags::smooth)
+        .def_readwrite("multiPort", &Segmenter::SegmenterFlags::multiPort)
+        .def_readwrite("ticksAsREEventNum", &Segmenter::SegmenterFlags::ticksAsREEventNum)
+        .def_readwrite("lbHdrVersion", &Segmenter::SegmenterFlags::lbHdrVersion)
         .def("getFromINI", &Segmenter::SegmenterFlags::getFromINI);
 
     // Constructor-simple
@@ -267,6 +273,7 @@ void init_e2sarDP_reassembler(py::module_ &m) {
         .def_readwrite("weight", &Reassembler::ReassemblerFlags::weight)
         .def_readwrite("min_factor", &Reassembler::ReassemblerFlags::min_factor)
         .def_readwrite("max_factor", &Reassembler::ReassemblerFlags::max_factor)
+        .def_readwrite("reportStats", &Reassembler::ReassemblerFlags::reportStats)
         .def("getFromINI", &Reassembler::ReassemblerFlags::getFromINI);
 
     // Constructor-simple